fix: move ConfirmModal folder to components folder
